splitting up fitting tests
